+++ cpp/arduino/AvrAdc.cpp
@@ -1,0 +1,4 @@
+#include "AvrAdc.h"
+
+// mock storage to allow access to ADCSRA
+unsigned char sfr_store;